+++ src/components/business_components/component.rs
@@ -1,0 +1,3 @@
+pub trait BusinessComponent {
+    async fn initialize_component(&mut self) {}
+}
