+++ src/components/business_components/mod.rs
@@ -1,3 +1,4 @@
+pub mod component;
 pub mod components;
 mod database;
 pub mod home;
